Fix scrolled-zoom (and broke preloading)
(Old svn revision: 3796)

+++ src/navigator.h
@@ -54,7 +54,7 @@ struct _RsttoNavigator
     GList                 *file_list;
     GList                 *file_iter;
     GList                 *history;
-    gint                   max_history;
+    gdouble                max_history;
 
     gboolean               preload;
